backend/pyproject.toml
[project]
name = "quant-platform-backend"
version = "0.1.0"
description = "Quant platform backend service"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110.0",
  "gunicorn>=22.0.0",
  "uvicorn[standard]>=0.27.0",
  "duckdb>=0.10.0",
  "pandas>=2.2.0",
  "pyarrow>=14.0.0",
  "tushare>=1.4.0",
  "pydantic-settings>=2.2.1",
  "apscheduler>=3.10.4",
  "tqdm>=4.66.0",
  "pymongo>=4.7.0",
  "passlib[bcrypt]>=1.7.4",
  "python-jose>=3.3.0",
  "bcrypt<4",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
